Struct pusher::Channel

pub struct Channel {
    pub occupied: Option<bool>,
    pub user_count: Option<i32>,
    pub subscription_count: Option<i32>,
}

This represents the data received upon querying the state of a Pusher channel.

Fields

occupied

Is the channel occupied?

user_count

The number of users presently subscribed to the channel

subscription_count

For accounts with subscription-counting enabled, the number of users currently subscribed to the channel.